store duplicate functions in sos/docker/client.py and sos/singularitry/client.py inside of a single file called imports.py
